Use the name of the end state rather than the name of the transition
In JIRA, you can name a state transition whatever you want -- the default name
is the name of the end state, but you can change that. It looks like the API
does expose the name of the end state, so we should key off of that, in case
the transition has a different name from the end state.

Note that this assumes that there will never be more than one transition ending
at the same state. If that happens, one of them will be clobbered. (It's not
clear that we have a good way of dealing with that even before this change.)