Be resilient in the face of GitHub flakiness
GitHub would send us a pull request event, and then serve us a 404 when
we tried to list the comments on the pull request.  This commit adds
retries for some GET requests.

Also, a way to simulate that flaky behavior in FakeGitHub, and a way to
enable that during testing.