When updating Jira with GitHub activity, it's OK if it fails.
When changing the title of an OSPR issue to Blended, we delete a Jira
issue and make a new one.  At the same time, the bot tries to update
Jira tickets with the latest activity.  It can get the id of the deleted
issue and then try to update it, resulting in a 404.  Don't let that
failure stop the process.