refactor(jira): remove inject_jira decorator
as it appears to be unnecessarily over-engineered.

The idea is clever: to automatically patch calls to inject a `jira`
client into calls if it was omitted.
But clever has a habit of catching up with you...

Nearly every single call already does pass this parameter along; in the
two locations where it is _not_ (at the top of the call stack), we can
just perform an explicit lookup and pass it in ourselves. Everywhere
else, we're applying this decorator to inject a variable that we expect
to be passed along anyway.

I struggle to see the utility...
Worse, it makes the debugging even more difficult.

This change should also avoid connecting to JIRA at import time,
else it tries to connect when we import from
`openedx_webhooks.lib.jira.client`.