feat: Point the openedx_webhooks bot to the new data repo.
As a part of the transition from edX to tCRIL, the repo-tools-data repo did
not get transferred.  A copy of the data needed by this application was made
and put into the openedx-webhooks-data repo in the openedx org.

This change points the app to load data from that repo instead.  Operationally,
that repo has the credentials for the edx-webhooks GitHub user and so that user
has also been given access to the new openedx-webhooks-data repo.