test: Fix repo rescan tests.
Previously this test relied on a behavior of the arrow library that when it was given
`None` as input, it would produce a current UTC time object.  Since the uprade of arrow,
it now produces an exception in this case.

Our test case didn't provide a closed_at time for closed PRs which meant the closed_at
time defaulted to `None`.  This is a situation we wouldn't run into in the real world
so fixing just the test is fine.