refactor: Update settings to work with Celery 5
Based on https://docs.celeryq.dev/en/stable/whatsnew-5.2.html#id5

Going through the changes needed as a part of the upgrade from celery 4 to
celery 5.  We update the CLI invocation but we're gonna skip the celery name
updates for now.

We're skipping the renaming because the old names will continue to work until
celery 6 and the new lowercase names make it so that the way we update the config
right now breaks.  Flask ignores lowercase settings so we need to nest them and
pick them up slightly differently.  We'll make those changes separately.

NOTE: CELERY_EAGER_PROPAGATES_EXCEPTIONS looks like it got renamed in the
celery 3 -> celery 4 upgrade but the update got missed so we'll just update
that one to the relevant celery 4 setting.