feat: conditional use of Jira
The plans for the bot are that some projects will want Jira tickets
created, and some will not.  This commit introduces a decision point for
which Jira projects to use, if any.  Right now, the code still always
makes Jira tickets in the same places it used to.

But the tests are generalized to run both with and without Jira access.