build: Actually use the common constraints file.
Previously we were downloading and checking in the
common_constraints.txt file as a part of the `make upgrade` process but
it was not actually being referenced by any of the other requirements
file.

Change it so that we reference the remote common_constraints.txt file
directly in the constraints.txt file.