feat: internal-ghorgs setting in orgs.yaml
A pull request is now considered internal if the author is internal to
the pull request's GitHub organization.  This is determined by the
author's institution's "internal-ghorgs" setting, a list of strings.

This code still reads "internal" if present, so that we can deploy the
new code before we update the data files.  Once the data files are
updated, we can remove the code using the "internal" setting.