feat: support running CLA check in merge queues
When a repo within the openedx org is configured to use merge queues, a CLA check response is requested for each PR in the queue. This check was never being run, leading to the PR being removed from the queue. This updates the hook to support the enqueued action type for PRs to ensure the check is run for PRs in the queue.