feat: Use a TLS redis url if it's available.
This is in preparation of an upgrade of the Redis server in production.

In production we use a premium redis server but it's running a fairly
old version of redis(4.x).  Heroku is about to drop support and force us
to use redis 7.x instead.  Before we do that, we want to make sure that
the server won't have any issues connecting to redis over a TLS
connection.  So we want to test it in staging first.

Unfortunately, n the staging environment we use the heroku redis mini
product.  This product allows for connecting over either a TLS
connection or a non-tls connection and provides a second environment
variable (REDIS_TLS_URL) if you want to connect to redis over TLS.  When
we upgrade production it will not provide this URL and the REDIS_URL
value will be secure by default.

So to validate we won't have issues in production, we force staging to
use the REDIS_TLS_URL if it's available but fallback to the REDIS_URL if
it is not available.