feat!: Remove rq.
It looks like we're not actually using rq anywhere in the current
implementation so let's get rid of it.  The only reference left to it is
in a comment for authenticating to Jira so get rid of it.

Any future delayed work should be done using celery tasks.

BREAKING CHANGE: You can't enqueue work to RQ workers anymore.