fix: Add celery TLS option.
Celery requires more options be set if using a TLS redis URL.  We add it
here instead of doing so in the environment variable because the setting
comes from heroku and any updates we make there would get over-written.

Unfortunately we can't verify the TLS cert because the "Heroku Data for
Redis" product uses self-signed certificates.