fix: Don't mention the owner directly in the welcome message.
We had backticks before and we removed them when we re-worded this.
However, we should put them back in.  Currently the maintainers get
mentioned and there fore start recieving notifications on all PRs
whether or not they are ready for their review.  This is causing a bit
of churn and waste as people have to check more PRs than necessary
especially since many PRs have known reviewers from the same team or via
FCs.