fix: Treat github usernames in a case-insensitively
Github treats them in a case insensitive manner but users can still
choose the casing of their name.  Our code for looking up users
previously was case-sensitive which was causing failures when someone
changed the casing in their github username.

This change updates the code so that the name we get associated with a
PR and the name we look up in the salesforce exports are both lowercased
before we do any comparisons.