temp: Add more logging to understand a failure case.
In production we're seeing that sometimes the state comes through as
"merged" but the "merged_at" field is missing.  This field should be a
required field on the pull_request webhoop so it's unclear why this is
happing.

For now just log the whole pr object so we can more easily understand
what's happening.