fix: Check for `merged_at` in two places.
GitHub does not make it clear what the type of the webhook data will be
and in some events the `merged_at` field is at the top-level and in some
hooks it's in a `pull_request` sub-dictionary.  The docs on this are not
super clear so just check both places for now so that we don't miss out
on the data.  This also fails with a warning instead of trowing an
exception which should make the rest of the code a bit more resilient.